requirements.txt
serpapi
fastapi
uvicorn
python-dotenv